+++ bindings/nostr-sdk-js/src/pool/mod.rs
@@ -258,7 +258,7 @@ impl JsRelayPool {
             .into())
     }
 
-    /// Subscribe to filters to all connected relays
+    /// Subscribe to filters to relays with `READ` flag.
     ///
     /// ### Auto-closing subscription
     ///
@@ -281,7 +281,7 @@ impl JsRelayPool {
             .map(|o| o.into())
     }
 
-    /// Subscribe with custom subscription ID to all connected relays
+    /// Subscribe with custom subscription ID to relays with `READ` flag.
     ///
     /// ### Auto-closing subscription
     ///